Cap reload times according to remaining ammo

+++ src/player.h
@@ -773,9 +773,10 @@ class player : public Character, public JsonSerializer, public JsonDeserializer
         /**
          * Calculate (but do not deduct) the number of moves required to reload an item with specified quantity of ammo
          * @param ammo either ammo or magazine to use when reloading the item
-         * @param qty units of ammo defaulting to remaining capacity (or 1 if RELOAD_ONE). Ignored if reloading using a magazine.
+         * @param qty maximum units of ammo to reload capped by remaining capacity. Defaults to remaining capacity
+         * (or 1 if RELOAD_ONE). Ignored if reloading using a magazine.
          */
-        int item_reload_cost( const item& it, const item& ammo, int qty = -1 ) const;
+        int item_reload_cost( const item& it, const item& ammo, long qty = -1 ) const;
 
         /** Wear item; returns false on fail. If interactive is false, don't alert the player or drain moves on completion. */
         bool wear(int pos, bool interactive = true);
